.legacy/requirements.txt
boto3==1.6.9
Flask==1.0
Flask-Login==0.4.1
Flask-OAuth==0.12
Flask-SocketIO==2.9.4
eventlet==0.22.1
psycopg2==2.7.4
python-engineio==2.0.3
python-socketio==1.9.0
requests_oauthlib==0.8.0
SQLAlchemy==1.3.13
yappi==0.98
Werkzeug==0.14.1
